add in more info and simplify backend model
