+++ BTCPayServer/Views/AppsPublic/ViewCrowdfund.cshtml
@@ -23,7 +23,8 @@
         <script type="text/javascript">
             var srvModel = @Html.Raw(Json.Serialize(Model));
         </script>
-        <bundle name="wwwroot/bundles/crowdfund-bundle.min.js"></bundle>
+        <bundle name="wwwroot/bundles/crowdfund-bundle-1.min.js"></bundle>
+        <bundle name="wwwroot/bundles/crowdfund-bundle-2.min.js"></bundle>
         <bundle name="wwwroot/bundles/crowdfund-bundle.min.css"></bundle>
     }
     @if (!string.IsNullOrEmpty(Model.EmbeddedCSS))
